fix!(ListaDeseos): Corregir tipo de dato llave

+++ backend/src/main/java/com/debede/backend/repository/ListaDeseos_ProductoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ListaDeseos_ProductoRepository extends JpaRepository<ListaDeseos_Producto, Long> {
+public interface ListaDeseos_ProductoRepository extends JpaRepository<ListaDeseos_Producto, Integer> {
 
 }
